feat(utils): move getIngredientsPrice to utils
