+++ src/utils/get-ingredients-price/index.ts
@@ -1,2 +1,1 @@
 export { getIngredientsPrice } from './get-ingredients-price';
-export { getTextDay } from './get-text-day';
